Module 2/FPTHelpDesk/backend/pyproject.toml
[project]
name = "fpthelpdesk"
version = "0.1.0"
description = "FPT HelpDesk — FastAPI + LangGraph multi-agent system"
requires-python = ">=3.11"
dependencies = [
    # FastAPI & Server
    "fastapi>=0.104.0",
    "uvicorn[standard]>=0.24.0",
    "python-multipart>=0.0.6",

    # Database
    "sqlalchemy[asyncio]>=2.0.0",
    "asyncpg>=0.29.0",
    "psycopg[binary]>=3.1.0",
    "pgvector>=0.3.0",

    # LangChain & LangGraph
    "langchain-openai>=0.1.0",
    "langchain-community>=0.2.0",
    "langchain-core>=0.3.0",
    "langchain-postgres>=0.0.12",
    "langgraph>=0.2.0",
    "langgraph-checkpoint>=2.0.0",
    "langgraph-checkpoint-postgres>=1.0.0",

    # Vector Store & Documents
    "pypdf>=4.0.0",
    "langchain-text-splitters>=0.2.0",

    # Search
    "langchain-tavily>=0.1.0",

    # Auth & Security
    "python-jose[cryptography]>=3.3.0",
    "passlib[bcrypt]>=1.7.4",
    "bcrypt>=3.2.0,<4.0.0",

    # Config
    "pydantic-settings>=2.0.0",
    "python-dotenv>=1.0.0",
]

[tool.uv]
# Xác định package là app (không phải library), không build wheel
package = false

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
